lib: utils: Add fdt_add_cpu_idle_states() helper function
Since the availability and latency properties of CPU idle states depend
on the specific SBI HSM implementation, it is appropriate that the idle
states are added to the devicetree by that implementation at runtime.

This helper function adds a platform-provided list of idle states to the
devicetree, using the SBI idle state binding. It makes some assumptions
for simplicity, but these could be relaxed if needed in the future.

Signed-off-by: Samuel Holland <samuel@sholland.org>